Move payment methods to partial class (file name TelegramBotClientPayment)
