+++ NetTelebot/NetTelebot.csproj
@@ -56,6 +56,7 @@
     <Compile Include="Result\FileInfoResult.cs" />
     <Compile Include="Result\IntegerResult.cs" />
     <Compile Include="Result\UserInfoResult.cs" />
+    <Compile Include="TelegramBotClientPayment.cs" />
     <Compile Include="TelegramBotClientUpdate.cs" />
     <Compile Include="Type\AudioInfo.cs" />
     <Compile Include="Result\BooleanResult.cs" />
